+++ example/gatsby-config.js
@@ -16,7 +16,15 @@ module.exports = {
     {
       resolve: "gatsby-transformer-remark",
       options: {
-        plugins: ["gatsby-remark-images", "gatsby-remark-prismjs"],
+        plugins: [
+          "gatsby-remark-images",
+          {
+            resolve: "gatsby-remark-prismjs",
+            options: {
+              noInlineHighlight: true,
+            },
+          },
+        ],
       },
     },
   ],