+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 -   Added: Script `gatsby-source-google-docs-token` for token generation
 -   Added: New documentation
 -   Added: Gatsby Cloud compatibility
+-   Added: crosslinks between documents
+-   Added: Footnotes support
 -   Updated: Simpler token behavior
 -   Deleted: `config` option